requirements.txt
# Core RL frameworks
gymnasium>=0.29.0
stable-baselines3>=2.2.0
sb3-contrib>=2.2.0

# Physics simulation
mujoco>=3.0.0
pybullet>=3.2.5

# Deep learning
torch>=2.0.0
tensorboard>=2.14.0

# Distributed training
ray[rllib]>=2.8.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0
seaborn>=0.12.0

# Data handling
h5py>=3.9.0
pyyaml>=6.0

# Logging and monitoring
wandb>=0.15.0
mlflow>=2.8.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
rich>=13.5.0
psutil>=5.9.0

# Optional: Advanced visualization
# pyvista>=0.42.0
# open3d>=0.17.0
# dash>=2.13.0
# streamlit>=1.26.0
